e2e: Move DeleteDiscoveredApps to disapp
Grouping related code in single file makes it easier to work with. More
work is needed to get rid of the crud.go.
