+++ e2e/deployers/crud.go
@@ -7,7 +7,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"os/exec"
 
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
@@ -411,38 +410,6 @@ func DeleteApplicationSet(ctx types.TestContext, a ApplicationSet) error {
 	}
 
 	log.Debugf("Deleted applicationset \"%s/%s\" in cluster %q", managementNamespace, name, ctx.Env().Hub.Name)
-
-	return nil
-}
-
-func DeleteDiscoveredApps(ctx types.TestContext, cluster types.Cluster, namespace string) error {
-	log := ctx.Logger()
-
-	tempDir, err := os.MkdirTemp("", "ramen-")
-	if err != nil {
-		return err
-	}
-
-	// Clean up by removing the temporary directory when done
-	defer os.RemoveAll(tempDir)
-
-	if err = CreateKustomizationFile(ctx, tempDir); err != nil {
-		return err
-	}
-
-	cmd := exec.Command("kubectl", "delete", "-k", tempDir, "-n", namespace,
-		"--kubeconfig", cluster.Kubeconfig, "--wait=false", "--ignore-not-found=true")
-
-	if out, err := cmd.Output(); err != nil {
-		if ee, ok := err.(*exec.ExitError); ok {
-			return fmt.Errorf("%w: stdout=%q stderr=%q", err, out, ee.Stderr)
-		}
-
-		return err
-	}
-
-	log.Debugf("Deleted discovered app \"%s/%s\" in cluster %q",
-		namespace, ctx.Workload().GetAppName(), cluster.Name)
 
 	return nil
 }
